services/blog_service/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy[asyncio]
asyncmy
python-dotenv
python-multipart
redis
passlib[bcrypt]